threat_watcher/requirements.txt
pymongo
python-dotenv
requests

